DeepZINB {DeepZINBLDA}	R Documentation

DeepZINBLDA

Description

Genetic cell classification is extremely important no matter in medical or biological fields. We design an antoencoder network to learn the parameters of the model, and use the zero-inflated negative binomial distribution as the loss function for better feature selection. Then we put the trained parameters into the zero-inflated negative binomial linear discriminant model, which based on the Bayesian model. Finally, the category of the sample is determined according to the discriminant score.

Usage

DeepZINB(X_train, X_test, y_train, y_test, batch_size, path_py)

Arguments

X_train
Matrix or data.The input training sample.

X_test
Matrix or data.The input test sample.

y_train
A vector shows the class of the input sample

y_test
A vector shows the class of the output sample

batch_size
Number of training batches of the neural network.

path_py
The path of python.exe.

Value

Misclassification of DeepZINBLDA

Examples

##simulation data
library(reticulate)
library(PoiClaClu)
library(limma)
library(sSeq)
library(edgeR)
library(DeepZINBLDA)
library(stringr)
X_train<-data(simdata_xtr)
X_train<-simdata_xtr
y_train<-data(simdata_ytr)
y_train<-simdata_ytr
X_test<-data(simdata_xte)
X_test<-simdata_xte
y_test<-data(simdata_yte)
y_test<-simdata_yte
batch_size=64
path_py="F:/python/Anaconda3/python.exe"
Misclassification<-DeepZINB(X_train,X_test,y_train,y_test,batch_size,path_py)

[Package DeepZINBLDA version 0.0.9000 Index]